move defaults to precompile folder and use existing builder pattern
