+++ x/vm/keeper/keeper_test.go
@@ -96,6 +96,7 @@ func (suite *KeeperTestSuite) SetupTest() {
 		suite.consensusKeeper,
 		suite.erc20Keeper,
 		"",
+		nil,
 	)
 }
