+++ x/vm/keeper/precompiles.go
@@ -1,26 +1,9 @@
 package keeper
 
 import (
-	"cosmossdk.io/core/address"
-	"fmt"
-	addresscodec "github.com/cosmos/cosmos-sdk/codec/address"
-	distributionkeeper "github.com/cosmos/cosmos-sdk/x/distribution/keeper"
-	govkeeper "github.com/cosmos/cosmos-sdk/x/gov/keeper"
-	slashingkeeper "github.com/cosmos/cosmos-sdk/x/slashing/keeper"
-	stakingkeeper "github.com/cosmos/cosmos-sdk/x/staking/keeper"
-	bankprecompile "github.com/cosmos/evm/precompiles/bank"
-	"github.com/cosmos/evm/precompiles/bech32"
-	distprecompile "github.com/cosmos/evm/precompiles/distribution"
-	govprecompile "github.com/cosmos/evm/precompiles/gov"
-	ics20precompile "github.com/cosmos/evm/precompiles/ics20"
-	"github.com/cosmos/evm/precompiles/p256"
-	slashingprecompile "github.com/cosmos/evm/precompiles/slashing"
-	stakingprecompile "github.com/cosmos/evm/precompiles/staking"
+	"github.com/cosmos/evm/x/vm/types"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/vm"
-	"maps"
-
-	"github.com/cosmos/evm/x/vm/types"
 
 	sdktypes "github.com/cosmos/cosmos-sdk/types"
 )
@@ -81,139 +64,4 @@ func (k *Keeper) GetPrecompilesCallHook(ctx sdktypes.Context) types.CallHook {
 
 		return nil
 	}
-}
-
-// Optionals define some optional params that can be applied to _some_ precompiles.
-// Extend this struct, add a sane default to defaultOptionals, and an Option function to provide users with a non-breaking
-// way to provide custom args to certain precompiles.
-type Optionals struct {
-	AddressCodec       address.Codec // used by gov/staking
-	ValidatorAddrCodec address.Codec // used by slashing
-	ConsensusAddrCodec address.Codec // used by slashing
-}
-
-func defaultOptionals() Optionals {
-	return Optionals{
-		AddressCodec:       addresscodec.NewBech32Codec(sdktypes.GetConfig().GetBech32AccountAddrPrefix()),
-		ValidatorAddrCodec: addresscodec.NewBech32Codec(sdktypes.GetConfig().GetBech32ValidatorAddrPrefix()),
-		ConsensusAddrCodec: addresscodec.NewBech32Codec(sdktypes.GetConfig().GetBech32ConsensusAddrPrefix()),
-	}
-}
-
-type Option func(opts *Optionals)
-
-func WithAddressCodec(codec address.Codec) Option {
-	return func(opts *Optionals) {
-		opts.AddressCodec = codec
-	}
-}
-
-func WithValidatorAddrCodec(codec address.Codec) Option {
-	return func(opts *Optionals) {
-		opts.ValidatorAddrCodec = codec
-	}
-}
-
-func WithConsensusAddrCodec(codec address.Codec) Option {
-	return func(opts *Optionals) {
-		opts.ConsensusAddrCodec = codec
-	}
-}
-
-const bech32PrecompileBaseGas = 6_000
-
-// DefaultStaticPrecompiles returns the list of all available static precompiled contracts from Cosmos EVM.
-//
-// NOTE: this should only be used during initialization of the Keeper.
-func (k *Keeper) DefaultStaticPrecompiles(
-	opts ...Option,
-) map[common.Address]vm.PrecompiledContract {
-	options := defaultOptionals()
-	for _, opt := range opts {
-		opt(&options)
-	}
-	// Clone the mapping from the latest EVM fork.
-	precompiles := maps.Clone(vm.PrecompiledContractsPrague)
-
-	// secp256r1 precompile as per EIP-7212
-	p256Precompile := &p256.Precompile{}
-
-	bech32Precompile, err := bech32.NewPrecompile(bech32PrecompileBaseGas)
-	if err != nil {
-		panic(fmt.Errorf("failed to instantiate bech32 precompile: %w", err))
-	}
-
-	stakingPrecompile, err := stakingprecompile.NewPrecompile(
-		k.stakingKeeper,
-		stakingkeeper.NewMsgServerImpl(k.stakingKeeper),
-		stakingkeeper.NewQuerier(k.stakingKeeper),
-		k.bankWrapper,
-		options.AddressCodec,
-	)
-	if err != nil {
-		panic(fmt.Errorf("failed to instantiate staking precompile: %w", err))
-	}
-
-	distributionPrecompile, err := distprecompile.NewPrecompile(
-		k.distrKeeper,
-		distributionkeeper.NewMsgServerImpl(*k.distrKeeper),
-		distributionkeeper.NewQuerier(*k.distrKeeper),
-		k.stakingKeeper,
-		k.bankWrapper,
-		options.AddressCodec,
-	)
-	if err != nil {
-		panic(fmt.Errorf("failed to instantiate distribution precompile: %w", err))
-	}
-
-	ibcTransferPrecompile, err := ics20precompile.NewPrecompile(
-		k.bankWrapper,
-		k.stakingKeeper,
-		k.transferKeeper,
-		k.channelKeeper,
-	)
-	if err != nil {
-		panic(fmt.Errorf("failed to instantiate ICS20 precompile: %w", err))
-	}
-
-	bankPrecompile, err := bankprecompile.NewPrecompile(k.bankWrapper, k.erc20Keeper)
-	if err != nil {
-		panic(fmt.Errorf("failed to instantiate bank precompile: %w", err))
-	}
-
-	govPrecompile, err := govprecompile.NewPrecompile(
-		govkeeper.NewMsgServerImpl(k.govKeeper),
-		govkeeper.NewQueryServer(k.govKeeper),
-		k.bankWrapper,
-		k.cdc,
-		options.AddressCodec,
-	)
-	if err != nil {
-		panic(fmt.Errorf("failed to instantiate gov precompile: %w", err))
-	}
-
-	slashingPrecompile, err := slashingprecompile.NewPrecompile(
-		k.slashingKeeper,
-		slashingkeeper.NewMsgServerImpl(*k.slashingKeeper),
-		k.bankWrapper,
-		options.ValidatorAddrCodec,
-		options.ConsensusAddrCodec,
-	)
-	if err != nil {
-		panic(fmt.Errorf("failed to instantiate slashing precompile: %w", err))
-	}
-
-	// Stateless precompiles
-	precompiles[bech32Precompile.Address()] = bech32Precompile
-	precompiles[p256Precompile.Address()] = p256Precompile
-
-	// Stateful precompiles
-	precompiles[stakingPrecompile.Address()] = stakingPrecompile
-	precompiles[distributionPrecompile.Address()] = distributionPrecompile
-	precompiles[ibcTransferPrecompile.Address()] = ibcTransferPrecompile
-	precompiles[bankPrecompile.Address()] = bankPrecompile
-	precompiles[govPrecompile.Address()] = govPrecompile
-	precompiles[slashingPrecompile.Address()] = slashingPrecompile
-
-	return precompiles
 }
